Related work items: #11449211, #11750336, #11508867, #11739564, #11545712, #11695968, #11801503, #11564233, #11709274, #11549839, #11437652

+++ src/interactivity/onecore/BgfxEngine.cpp
@@ -156,7 +156,7 @@ HRESULT BgfxEngine::PaintBufferLine(PCWCHAR const pwsLine, size_t const cchLine,
 
     USHORT PaletteEntry = GetColorPaletteEntry(_currentForegroundColor);
 
-    for (size_t i = 0 ; i < cchLine && i < _displayWidth ; i++)
+    for (size_t i = 0 ; i < cchLine && i < (size_t)_displayWidth ; i++)
     {
         NewRun[coord.X + i].Character = pwsLine[i];
         NewRun[coord.X + i].Atribute = PaletteEntry;
